+++ src/Symfony/Component/HttpFoundation/CHANGELOG.md
@@ -8,7 +8,8 @@ CHANGELOG
    disabling `Range` and `Content-Length` handling, switching to chunked encoding instead
  * added the `Cookie::fromString()` method that allows to create a cookie from a
    raw header string
-
+ * PdoSessionHandler: [BC BREAK] removed the `lifetime` and `timestamp` columns in favor of `expiry` column
+ 
 3.1.0
 -----